+++ src/bp/ui-admin/src/utils/toaster.tsx
@@ -15,26 +15,3 @@ export const toastFailure = message =>
     intent: Intent.DANGER,
     timeout: 3000
   })
-
-export const toastError = error => {
-  const errorCode = _.get(error, 'response.data.errorCode') || _.get(error, 'errorCode')
-  const details = _.get(error, 'response.data.message') || _.get(error, 'message')
-  const docs = _.get(error, 'response.data.docs') || _.get(error, 'docs')
-
-  let message = (
-    <span>
-      {errorCode && <span>[{errorCode}]</span>} {details}{' '}
-      {docs && (
-        <a href={docs} target="_blank">
-          More informations
-        </a>
-      )}
-    </span>
-  )
-
-  if (!errorCode && !message) {
-    message = <span>Something wrong happened. Please try again later.</span>
-  }
-
-  toastFailure(message)
-}
